* master: (249 commits)
  Gradle: Try to use Bullet Native for the SDK by moving jbullet to optlibs and bullet-native to corelibs
  Gradle Build: Add convenience task `run` for Windows Systems which cannot invoke ant from the command line. (And each project should have a run task, anyway)
  Fix source version
  iOS: Set openGL Context in AppDelegate to prevent Crashes.
  Travis: Try to reduce disk space usage
  Build: Remove HW Skinning Patch as it has been merged into the engine
  Build: Try to fix regular commits erroring due to a faulty engine-version-name
  Updated Bloom filter wizard's name
  Fixes #91 iOS: Automatically download the prebuilt Avian Version to make iOS Deployment working again.
  Fixes #75 - DarkMonkey: We can now change the Diff-Colors and the User can adjust them manually using the Fonts & Colors Dialog. Also see 1e92c2f
  Fixes #88 - TextureBrowser: Release the old Texture when setting the path to a new one.
  Update XBuf Dependency to 0.9.1 (ProtoBuf 3)
  Fix Regression with fixPlatformIndependentBuild Gradle Task.
  iOS: Update Generated Files
  iOS: Update native Image-Loading Code to prevent Crashes
  iOS: Fix crash in GLES Native Code due to changed Method Signatures
  iOS Harness: Use newer GLRenderer over old Implementation.
  Fix Constructor of ProjectExtensionManager
  iOS: Build x86_64 instead of i386, update jar-excludes (new naming scheme of jme), ignore "ios-base" Library since it doesn't seem to be used anymore.
  iOS: Also package the x86_64 avian version
  ...

Conflicts: (solved)
	branding/core/core.jar/org/netbeans/core/startup/Bundle.properties
	branding/modules/org-netbeans-core-windows.jar/org/netbeans/core/windows/view/ui/Bundle.properties
	build.gradle
	jme3-scenecomposer/src/com/jme3/gde/scenecomposer/SceneComposerToolController.java
	jme3-scenecomposer/src/com/jme3/gde/scenecomposer/gizmo/light/LightGizmoFactory.java
	nbproject/project.properties

+++ jme3-terrain-editor/src/com/jme3/gde/terraineditor/TerrainEditorController.java
@@ -1089,12 +1089,12 @@ public class TerrainEditorController implements NodeListener {
 
             if (enabled) {
                 for (int i=0; i<getNumUsedTextures(); i++) {
-                    float scale = 1f/(float)(size/getTextureScale(i));
+                    float scale = 1f/(size/getTextureScale(i));
                     setTextureScale(i, scale);
                 }
             } else {
                 for (int i=0; i<getNumUsedTextures(); i++) {
-                    float scale = (float)(size*getTextureScale(i));
+                    float scale = (size*getTextureScale(i));
                     setTextureScale(i, scale);
                 }
             }
